requirements.txt
numpy==1.18.1
opencv-contrib-python==3.4.5.20
python-vlc==3.0.7110
